// src/features/products/components/form/ProductFormGeneral.types.ts
export type ProductFormLanguage = {
  code: string;
  name: string;
};

export type ProductFormGeneralWatchedValues = {
  nameEn: unknown;
  namePl: unknown;
  nameDe: unknown;
  descEn: unknown;
  descPl: unknown;
  descDe: unknown;
  sku: unknown;
  weight: unknown;
  sizeLength: unknown;
  sizeWidth: unknown;
  fieldLength: unknown;
  supplierName: unknown;
  supplierLink: unknown;
  priceComment: unknown;
  price: unknown;
  stock: unknown;
};

export type ProductFormGeneralDisplayValues = Record<string, unknown>;